third_party/WebKit/LayoutTests/http/tests/websocket/receive-pong-with-payload.html

<!DOCTYPE HTML>
<script src="/js-test-resources/js-test.js"></script>
<script>
description("Test whether WebSocket correctly ignores a pong message with a payload");

window.jsTestIsAsync = true;

var url = "ws://localhost:8880/receive-pong-with-payload";
var ws = new WebSocket(url);

var message;
ws.onmessage = function(event)
{
    message = event.data;
    shouldBeEqualToString("message", "sent pong");
    ws.close();
};

var wasClean;
ws.onclose = function(event)
{
    wasClean = event.wasClean;
    shouldBeTrue("wasClean");
    finishJSTest();
};

</script>
