third_party/blink/web_tests/http/tests/websocket/receive-pong-with-payload-expected.txt

Test whether WebSocket correctly ignores a pong message with a payload

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS message is "sent pong"
PASS wasClean is true
PASS successfullyParsed is true

TEST COMPLETE
